project: rename to Insulate
