+++ src/insulate/__init__.py
@@ -1,0 +1,3 @@
+from insulate.registry import prepare_environment
+
+__version__ = "0.1.0"
