+++ src/insulate/backends/__init__.py
@@ -1,0 +1,1 @@
+from insulate.backends._base import *
