+++ src/insulate/backends/connections/__init__.py
@@ -1,4 +1,4 @@
-from isolate.backends.connections.ipc import (
+from insulate.backends.connections.ipc import (
     ExtendedPythonIPC,
     IsolatedProcessConnection,
     PythonIPC,
