+++ src/insulate/backends/connections/ipc/__init__.py
@@ -1,4 +1,4 @@
-from isolate.backends.connections.ipc._base import (
+from insulate.backends.connections.ipc._base import (
     ExtendedPythonIPC,
     IsolatedProcessConnection,
     PythonIPC,
