+++ src/insulate/backends/context.py
@@ -8,7 +8,7 @@ from typing import TYPE_CHECKING, Callable, Dict, NewType, Optional
 from platformdirs import user_cache_dir
 
 if TYPE_CHECKING:
-    from isolate.backends import BaseEnvironment
+    from insulate.backends import BaseEnvironment
 
 
 class LogSource(str, Enum):
@@ -97,7 +97,7 @@ ContextType = NewType("ContextType", _Context)
 GLOBAL_CONTEXT = ContextType(
     _Context(
         _base_cache_dir=Path(
-            user_cache_dir("isolate", "isolate"),
+            user_cache_dir("insulate", "insulate"),
         )
     )
 )
