+++ src/insulate/server/__init__.py
@@ -1,0 +1,1 @@
+from insulate.server._server import app
