--- src/isolate/__init__.py
@@ -1,3 +1,0 @@
-from isolate.registry import prepare_environment
-
-__version__ = "0.1.0"
