--- src/isolate/backends/__init__.py
@@ -1,1 +1,0 @@
-from isolate.backends._base import *
