--- src/isolate/server/__init__.py
@@ -1,1 +1,0 @@
-from isolate.server._server import app
